refact: move and renames behavior class plus cleanups
