+++ src/behaviors/BlocksLoaderBehavior.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace madebyraygun\blockloader\web\twig;
+namespace madebyraygun\blockloader\behaviors;
 
 use Craft;
 use yii\base\Behavior;
 use madebyraygun\blockloader\base\BlocksProvider;
 use Illuminate\Support\Collection;
 
-class BlocksLoader extends Behavior
+class BlocksLoaderBehavior extends Behavior
 {
     public function blocksFromField(string $fieldHandle): Collection
     {
